requirements.txt
streamlit
requests
instaloader
pillow
glob